refactor(codegen): carve runtime-ABI declaration interning out of llvm.rs
Relocate `intern_runtime_decl` (the runtime-ABI signature table), the
catalog-FFI declaration helpers, extern-record return classification,
and the stdlib-catalog / extern predeclaration from llvm.rs into the
existing `crate::runtime_abi` module, alongside the runtime-call
lowering it already owns. Callers in llvm/arith/suspend/thunks import
the entry points from `crate::runtime_abi`.

A pre-existing `bytes_triple_llvm_ty` name collision (a `&Context`
variant in llvm, a `&FnCtx` variant in runtime_abi) is disambiguated at
the two moved call sites by qualifying `crate::llvm::bytes_triple_llvm_ty`
so the interning path resolves to the same helper as before.

Pure relocation: every emitted `.ll` is byte-identical (ll-diff verified
byte-identical native + wasm32). No IR change.
