+++ hew-codegen-rs/src/thunks.rs
@@ -32,6 +32,7 @@ use hew_types::{BuiltinType, ResolvedTy};
 
 #[allow(unused_imports)]
 use crate::llvm::*;
+use crate::runtime_abi::intern_runtime_decl;
 
 fn task_wrapper_name(callee_symbol: &str) -> String {
     format!("__hew_task_wrapper_{}", sanitize_symbol(callee_symbol))
